Global Database Attribute
At the top of app.py below the app variable create a function called open_connection.

In the body of the open_connection function use the built_in getattr() function to get the '_connection' attribute from the g object, and set the default to None. Assign the return value of the getattr function to connection.