Query Database Function
Let’s create a function to make it easier to query the database.

Below the open_connection function in app.py create a function called execute_sql.

In the body of execute_sql create a variable called connection. Assign this variable the return value of a call to the newly created open_connection function.