1. Job Template
We need a template to display an individual job. Create a new file called job.html in the template folder.

In the file use an extends template tag to inherit layout.html.

After the extends tag add a template block called content. In the block call the show_job macro passing in job. Note: Use the {{}} for the macro call.