3. Job Route Decorator
We only need one job from the database, so we will use the execute_sql function passing in a query with a where clause. Before we do that, we'll need a job_id for the where clause, and we are going to get this from the URL.

Still in app.py, add a route decorator with the URL path /job/<job_id> to the job function.