4. Job Route Parameter
To use the job_id we received from the URL, we need to pass it to the job function. Add job_id to the parameter list of the job function.

Now that the job function has been created and has parameters, open _macros.html. Find <a> tag with the content: {{ job['title'] }}. Add an href with the value {{ url_for('job', job_id=job['id']) }}.