1. Employer Template
To display an employer, create a new file called employer.html in the templates folder. Open the templates.html file, find the appropriate block of HTML and copy and paste it to employer.html.

To the top of the file inherit from the layout.html template by using an extends template tag.