6. Employer Route
The template we have just built needs access to employer, job, and review data. Let's create a new function in app.py called employer.

Add a route decorator with a URL pattern of /employer/<employer_id>.

In the body return a call to the render_template function passing in the employer.html template.

Open _macros.html and find <a> tag with the content: {{ job['employer_name'] }}. Add an href with the value {{ url_for('employer', employer_id=job['employer_id']) }}.