8. Employer Route Employer Details
Still working with the employer function, add employer_id to the parameter list so that we have access to this value. Above the render_template function make a call to execute_sql and assign the return value to employer.

Pass the following arguments to execute_sql:

    SQL Query: 'SELECT * FROM employer WHERE id=?'
    List Literal: [employer_id]
    single=True: This will bring back only one result.

In the render_template function, pass a keyword argument of employer=employer.