9. Employer Route Employer Jobs
On the employer details page, we want to display all of the employers' jobs. In the employer function in app.py below the employer variable, add a call to the execute_sql function and assign the results to a variable called jobs. Pass the function two arguments:

    SQL Query: 'SELECT job.id, job.title, job.description, job.salary FROM job JOIN employer ON employer.id = job.employer_id WHERE employer.id = ?'
    List Literal: [employer_id]

In the render_template function, add another keyword argument of jobs=jobs