10. Employer Route Employer Review
Still in the employer function in app.py below the jobs query add a new query to get all review for the employer. Make a call to execute_sql and assign the return value to reviews. Pass in the arguments:

    SQL Query: 'SELECT review, rating, title, date, status FROM review JOIN employer ON employer.id = review.employer_id WHERE employer.id = ?'
    List Literal: [employer_id]

In the render_template function, add another keyword argument of reviews=reviews