2. Review Route
In app.py below the employer function create a new function called review. Add employer_id to the parameter list.

Add a route decorator with a URL pattern of /employer/<employer_id>/review. Also add a keyword argument methods set to a tuple with two values: 'GET' and 'POST'.

In the body of the function return the render_template function passing in the review.html template and a keyword argument of employer_id=employer_id.