5. Redirect to Employer Page
Redirect to Employer Page

At the end of the function redirect the user back to the employer page.

This can be done by using the redirect and url_for functions.

To start, import both of these functions from flask.

Next, in the if statement below the execute_sql call, return a call to redirect and pass in a call to the url_for function.

Pass the url_for function the route to redirect to, which is the employer route and a keyword argument of employer_id=employer_id.